refactor(ai_loop): centralize helper funcs
