+++ ai_loop/tests/test_logging.py
@@ -1,22 +1,17 @@
-import os
 import sys
-import importlib.util
 from pathlib import Path
+
+repo_root = Path(__file__).resolve().parents[2]
+sys.path.insert(0, str(repo_root))
+
+from ai_loop.utils_common import ensure_logs
 
 
 def test_ensure_logs_creates_files(tmp_path, monkeypatch):
-    monkeypatch.setenv('OPENAI_API_KEY', 'test')
     monkeypatch.chdir(tmp_path)
-    repo_root = Path(__file__).resolve().parents[2]
-    sys.path.insert(0, str(repo_root))
 
-    module_path = repo_root / 'ai_loop' / 'trendspire_autoloop.py'
-    spec = importlib.util.spec_from_file_location('trendspire_autoloop', module_path)
-    mod = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(mod)
+    ensure_logs('logs', 'costs.csv', 'mem')
 
-    mod.ensure_logs()
-
-    assert (tmp_path / mod.LOG_DIR).is_dir()
-    assert (tmp_path / mod.MEMORY_DIR).is_dir()
-    assert (tmp_path / mod.COST_LOG).is_file()
+    assert (tmp_path / 'logs').is_dir()
+    assert (tmp_path / 'mem').is_dir()
+    assert (tmp_path / 'costs.csv').is_file()
